Big change to how events are handled Recently due to time changes I made a change to publish events to the NSB transport from the domain handler.  Meaning after the domain wrote the events, it would read the events it wrote via the ES connection and Publish them with NSB This is not ideal - because if the domain instance crashes all the events it wrote between the last Publish and the last command will not get published to the bus.  It was a quick and dirty fix.
Now - consumers are back to reading from the eventstore.  When consumers start they take a list of all events they subscribe to (thanks NSB) and create an eventstore projection to project all those events into a stream for itself.  It then creates a persistent group connection to that stream and leverages the competing consumer feature in ES to scale out event consumers.
There are numerous benefits to this approach, here are some:
- consumers no longer subscribe to $all significantly reducing the events needed to be delivered
- if any instance crashes unexpectidly events dont go missing
- utilizing assembly version we can make versioned streams for very easy production upgrades
- less stress on NSB transport
- the weird domain/consumer/eventstore relationship triangle is seperated as it should be